+++ src/image_struct.h
@@ -1,12 +1,15 @@
 #ifndef image_struct_h
 #define image_struct_h
 
+#include <stdint.h>
+
+
 typedef struct ImageStruct {
 	unsigned int numColumns;
 	unsigned int height;
 	bool upsideDown;
 	unsigned char palette[16];
-	unsigned long int *pixels;
+	uint32_t *pixels;
 	unsigned char *mask;
 } ImageStruct;
